[widgets] Move DMA-transport related traits into their own file
