+++ src/main/scala/midas/core/FPGATop.scala
@@ -154,10 +154,6 @@ class FPGATop(simIoType: SimWrapperIO)(implicit p: Parameters) extends Module wi
           case _ =>
       }
       channels2Port(widget.io.hPort, endpoint(i)._2)
-    
-
-      if (widget.io.dma.nonEmpty)
-        dmaInfoBuffer += DmaInfo(widgetName, widget.io.dma.get, widget.io.dmaSize)
 
       widget match {
         case widget: HasDMA => dmaInfoBuffer += DmaInfo(widgetName, widget.dma, widget.dmaSize)
@@ -253,7 +249,4 @@ class FPGATop(simIoType: SimWrapperIO)(implicit p: Parameters) extends Module wi
     "DMA_WIDTH"      -> p(DMANastiKey).dataBits / 8,
     "DMA_SIZE"       -> log2Ceil(p(DMANastiKey).dataBits / 8)
   )
-
 }
-
-
